Move metrics tracker to shared utils for reuse
Relocated createMetricsTracker from src/plugin/src/reporting/metrics-tracker.js to src/shared/utils/metrics.js so the generic instrumentation helper lives alongside other shared utilities.

Updated plugin imports, reporting barrels, and tests to follow the new shared entry point.
