+++ src/shared/utils/index.js
@@ -8,6 +8,7 @@ export * from "./fs.js";
 export * from "./json.js";
 export * from "./line-breaks.js";
 export * from "./identifier-metadata.js";
+export * from "./metrics.js";
 export * from "./number.js";
 export * from "./numeric-options.js";
 export * from "./object.js";
